move s_layer_t to core, move layer impl to src
