+++ include/crystalnet/core/layer.hpp
@@ -8,12 +8,3 @@ struct s_layer_t {
     virtual s_node_t *operator()(s_model_ctx_t &, s_node_t *x) const = 0;
     virtual ~s_layer_t() {}
 };
-
-template <typename T> struct op_instance;
-
-template <typename T> struct unary_op_layer : s_layer_t {
-    s_node_t *operator()(s_model_ctx_t &ctx, s_node_t *x) const override
-    {
-        return ctx.make_operator(*op_instance<T>::get(), x);
-    }
-};
